+++ src/main/java/ru/octol1ttle/flightassistant/alerts/autoflight/AutopilotOffAlert.java
@@ -28,7 +28,7 @@ public class AutopilotOffAlert extends AbstractAlert {
     }
 
     @Override
-    public int renderECAM(TextRenderer textRenderer, DrawContext context, float x, float y, boolean highlight) {
+    public int renderECAM(TextRenderer textRenderer, DrawContext context, int x, int y, boolean highlight) {
         return HudComponent.drawHighlightedText(textRenderer, context, Text.translatable("alerts.flightassistant.autoflight.autopilot_off"), x, y,
                 FAConfig.hud().warningColor,
                 highlight && autoflight.apDisconnectionForced);
