+++ src/main/java/ru/octol1ttle/flightassistant/alerts/fault/IndicatorFaultAlert.java
@@ -31,7 +31,7 @@ public class IndicatorFaultAlert extends AbstractAlert {
     }
 
     @Override
-    public int renderECAM(TextRenderer textRenderer, DrawContext context, float x, float y, boolean highlight) {
+    public int renderECAM(TextRenderer textRenderer, DrawContext context, int x, int y, boolean highlight) {
         int i = 0;
         for (HudComponent component : renderer.faulted) {
             i += HudComponent.drawText(textRenderer, context, Text.translatable("alerts.flightassistant.fault.indicators." + component.getId()), x, y,
