+++ src/main/java/ru/octol1ttle/flightassistant/alerts/firework/FireworkUnsafeAlert.java
@@ -33,7 +33,7 @@ public class FireworkUnsafeAlert extends AbstractAlert {
     }
 
     @Override
-    public int renderECAM(TextRenderer textRenderer, DrawContext context, float x, float y, boolean highlight) {
+    public int renderECAM(TextRenderer textRenderer, DrawContext context, int x, int y, boolean highlight) {
         return HudComponent.drawHighlightedText(textRenderer, context, Text.translatable("alerts.flightassistant.firework.unsafe"), x, y,
                 FAConfig.hud().warningColor, highlight && data.isFlying);
     }
