+++ src/main/java/ru/octol1ttle/flightassistant/alerts/nav/gpws/ExcessiveDescentAlert.java
@@ -30,7 +30,7 @@ public class ExcessiveDescentAlert extends AbstractAlert {
     }
 
     @Override
-    public boolean renderCentered(TextRenderer textRenderer, DrawContext context, float x, float y, boolean highlight) {
+    public boolean renderCentered(TextRenderer textRenderer, DrawContext context, int x, int y, boolean highlight) {
         if (gpws.descentImpactTime <= PULL_UP_THRESHOLD) {
             HudComponent.drawHighlightedMiddleAlignedText(textRenderer, context, Text.translatable("alerts.flightassistant.pull_up"), x, y,
                     FAConfig.hud().warningColor, highlight);
