+++ src/main/java/ru/octol1ttle/flightassistant/alerts/other/StallAlert.java
@@ -30,7 +30,7 @@ public class StallAlert extends AbstractAlert {
     }
 
     @Override
-    public boolean renderCentered(TextRenderer textRenderer, DrawContext context, float x, float y, boolean highlight) {
+    public boolean renderCentered(TextRenderer textRenderer, DrawContext context, int x, int y, boolean highlight) {
         HudComponent.drawHighlightedMiddleAlignedText(textRenderer, context, Text.translatable("alerts.flightassistant.stall"), x, y, FAConfig.hud().warningColor, highlight);
 
         return true;
